==> scripts/Autres/20_select_data.Rmd <==
---
title: "Fish population trends in the rivers of Britanny"
subtitle: "Data preparation"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  bookdown::html_document2:
    toc: true
    toc_float: true
#  bookdown::word_document2
always_allow_html: true
params:
  depts: !r c(22, 29, 35, 56)
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Setup

## Load packages and functions

```{r}
# if necessary
# devtools::install_github("pascalirz/aspe")
library(aspe)
library(tidyverse)

# install custom functions
# source("../../ipr_trends/R/functions.R")
```

## Load data

```{r}
# retrieve most recent data file from repo 
rdata_tables <- misc_nom_dernier_fichier(
  repertoire = "../../../../projets/ASPE/raw_data/rdata",
  pattern = "^tables")
# load it
load(rdata_tables)
# test
```

Data file : `r str_replace(rdata_tables, "../../../../projets/ASPE/raw_data/rdata/", "")`

# Assembly of datasets

From the overall dataset of the Aspe database, only the monitoring networks are considered here.

-   Réseau de Contrôle de Surveillance (RCS)
-   Réseau Hydrobiologique Piscicole (RHP)

They are considered representative of the diversity of the rivers ecosystems in continental France.

Note that one site can contribute to more than both networks (true also for one survey).

The data before 2007 were discarded because there was too much turnover between sites and the field protocols were not settled yet.

## Join tables and filter

### Getting started

We start by creating elementary_samples "gateway" dataframe connecting tables ids + additional variables.

This dataframe is then filtered to keep only:

-   the surveys belonging to the representative networks
-   since 2007

The selected sampling networks (=aims) are:

```{r}
my_aims <- c("RCS – Réseau de Contrôle de Surveillance",
             "RHP – Réseau Hydrobiologique Piscicole")
```

Assembly and filtering of the dataset.

```{r}
elementary_samples <- mef_creer_passerelle() %>% 
  select(sta_id:pre_id) %>% # remove the unnecessary variables
  distinct() %>% # remove duplicates due to multiple fish batches per survey
  mef_ajouter_ope_date() %>% 
  mef_ajouter_objectif() %>% 
  mef_ajouter_ipr() %>% 
  filter(annee > 2006,
         obj_libelle %in% my_aims)

if(!is.null(params$depts)) {
  
  depts <- as.character(params$depts)
  
  elementary_samples <- elementary_samples %>% 
    mef_ajouter_dept() %>% 
    filter(dept %in% depts)
  
}
```

A few sites may have multiple surveys some years. We keep only the last (autumn) one.

```{r}
elementary_samples <- elementary_samples %>% 
  group_by(pop_id,
           annee,
           obj_libelle) %>% 
  filter(ope_date == max(ope_date)) %>%   # only last date for sites sampled more than once a year
  ungroup()
```

### Sampling points



```{r}
points <- elementary_samples %>% 
  select(pop_id,
         ope_id,
         obj_libelle) %>%
  distinct() %>% 
  group_by(pop_id,
           obj_libelle) %>%
  tally() %>%
  ungroup() %>%
  mutate(obj_libelle = str_sub(obj_libelle, start = 1, end = 3)) %>% 
  pivot_wider(names_from = obj_libelle,
              values_from = n)

# check -> must be 0
# points %>% count(pop_id) %>% filter(n > 1)
```


```{r}
n_surveys_by_point <- elementary_samples %>% 
  group_by(pop_id) %>% 
  summarise(n_ope = n_distinct(ope_id)) %>% 
  ungroup()

points <- points %>% 
  left_join(y = n_surveys_by_point)
```

A sample of the "points" dataframe:

```{r}
points %>% 
  arrange(-n_ope) %>% 
  DT::datatable()
```


```{r}
points %>% 
  ggplot(aes(x = n_ope)) +
  geom_bar()
```

**Removal of points with less than 8 surveys**

```{r}
my_pops <- points %>% 
  filter(n_ope > 7) %>% 
  pull(pop_id)
```


The total number of sampling sites is `r length(my_pops)`.

```{r}
points <- points %>% 
  filter(pop_id %in% my_pops)
```



### Surveys

```{r}
surveys <- elementary_samples %>% 
  filter(pop_id %in% my_pops) %>% 
  select(sta_id:ope_id,
         ope_date,
         annee) %>% 
  distinct()

surveys %>% 
  head() %>% 
  mutate(across(where(is.numeric), as.character)) %>% 
  flextable::flextable()
```

How many sites sampled by year?

```{r}
surveys %>% 
  group_by(annee) %>% 
  summarise(n_ope = n_distinct(pop_id)) %>% 
  ggplot(aes(x = annee, y = n_ope)) +
    geom_bar(stat = "identity") +
  labs(x = "Year",
       y = "Nb of sites surveyed")
```

## Elementary samples

```{r}
elementary_samples <- elementary_samples %>% 
  filter(pop_id %in% my_pops)
```

## Catches by species

```{r}
densities_ope <- elementary_samples %>% 
  select(-(obj_id:dept)) %>% 
  mef_ajouter_lots()
```

Recode some species names.

```{r}
densities_ope <- densities_ope %>% 
  mutate(esp_code_alternatif = case_when(
    esp_code_alternatif == "VAN" ~ "VAR",
    esp_code_alternatif == "CHA" ~ "CHF",
    TRUE ~ esp_code_alternatif
  ))
```

Aggregate by survey.

```{r}
densities_ope <- densities_ope %>% 
  select(esp_code_alternatif, everything()) %>% 
  select(-pre_id) %>% 
  group_by(across(esp_code_alternatif:annee)) %>% 
  summarise(n_indiv = sum(lop_effectif)) %>% 
  ungroup() %>% 
  mef_ajouter_surf_calc() %>% 
  mutate(density = n_indiv / ope_surface_calculee * 1000)
```


# Spatialise sampling points

Where are the sampling sites? 

The dataframe of the selected sampling points is transformed into a spatial object of class `sf`.

```{r}
points_geo <- point_prelevement %>%
  filter(pop_id %in% my_pops) %>%
  left_join(y = ref_type_projection,
            by = c("pop_typ_id" = "typ_id")) %>%
  geo_convertir_coords_df(
    var_x = pop_coordonnees_x,
    var_y = pop_coordonnees_y,
    var_id = pop_id,
    var_crs_initial = typ_code_epsg,
    crs_sortie = 4326
  ) %>%
  rename(x_wgs84 = X,
         y_wgs84 = Y) %>%
  sf::st_as_sf(coords = c("x_wgs84", "y_wgs84"),
               crs = 4326) %>%
  left_join(y = points) %>% 
  left_join(y = point_prelevement %>% 
              select(pop_id, sta_id = pop_sta_id)) %>% 
  mef_ajouter_libelle_site() %>% 
  select(pop_id,
         n_ope,
         pop_libelle)
```

The points sizes indicate the number of years the site has been surveyed.

```{r}
mapview::mapviewOptions(fgb = FALSE)

points_geo %>% 
  mapview::mapview(cex = "n_ope",
                   layer.name = "Nb of surveys",
                   min.rad = 4,
                   max.rad = 8,
                   row.numbers = FALSE,
                   popup = leafpop::popupTable(points_geo, feature.id = FALSE))
```


# Saving

```{r}
save(elementary_samples,
     surveys,
     densities_ope,
     my_pops,
     points,
     points_geo,
     file = "../processed_data/20_selected_data.RData")
```
